+++ src/main/java/com/scanoss/plugins/sonar/analyzers/ScanOSSAnalyzer.java
@@ -70,20 +70,16 @@ public class ScanOSSAnalyzer {
     private final Logger log = Loggers.get(this.getClass());
 
     /**
-     * SBOM identify file name
-     */
-    private final String sbomIdentify;
-
-
-    /**
-     * SBOM identify file name
-     */
-    private final String sbomIgnore;
-
-    /**
      * HPSM option
      */
     private final Boolean isHpsmEnabled;
+
+    /**
+     * SCANOSS SETTINGS
+     */
+    private final Boolean isScanossSettingsEnabled;
+
+    private final String scanossSettingsFilePath;
 
     /**
      * Analyzer constructor
@@ -93,15 +89,16 @@ public class ScanOSSAnalyzer {
      * @param key             SCANOSS API Access Key (optional)
      * @param customCertChain SCANOSS Custom Certificate Chain (optional)
      */
-    public ScanOSSAnalyzer(final File rootDir, String url, String key, String customCertChain, String sbomIdentify, String sbomIgnore, Boolean isHpsmEnabled) {
+    public ScanOSSAnalyzer(final File rootDir, String url, String key, String customCertChain, Boolean isHpsmEnabled,
+                           Boolean isScanossSettingsEnabled, String scanossSettingsFilePath) {
         super();
         this.rootDir = rootDir;
         this.url = url;
         this.key = key;
         this.customCertChain = customCertChain;
-        this.sbomIdentify = sbomIdentify;
-        this.sbomIgnore = sbomIgnore;
         this.isHpsmEnabled = isHpsmEnabled;
+        this.isScanossSettingsEnabled = isScanossSettingsEnabled;
+        this.scanossSettingsFilePath = scanossSettingsFilePath;
     }
 
     /**
@@ -112,7 +109,13 @@ public class ScanOSSAnalyzer {
     public ScanResult analyze(List<String> inputFilePaths)  {
         log.info("[SCANOSS] Starting scan process...");
         log.info("[SCANOSS] Plugin version: {} (SDK version: {})", PackageDetails.getVersion(), com.scanoss.utils.PackageDetails.getVersion());
-        ScanOSSScanner scanner = new ScanOSSScanner(this.url, this.key, this.customCertChain, this.sbomIdentify, this.sbomIgnore, this.isHpsmEnabled);
+        ScanOSSScanner scanner = new ScanOSSScanner(this.url,
+                this.key,
+                this.customCertChain,
+                this.isHpsmEnabled,
+                this.isScanossSettingsEnabled,
+                this.scanossSettingsFilePath
+        );
         List<String> output = scanner.runScan(rootDir.getPath(), inputFilePaths);
         if(output == null || output.isEmpty()){
             log.warn("[SCANOSS] Empty result");